app/requirements.txt
Django==3.2.25
psycopg2-binary==2.9.6
django-bootstrap5==23.1
backports.zoneinfo==0.2.1;python_version<"3.9"
djangorestframework
markdown==3.0.0
django-filter==1.0.1